refactor: move `get_unverified_jwt_claims` and `sd_jwt_to_jwt` to dedicated file
