+++ oid4vc-core/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod did;
 pub mod form_urlencoded;
+pub mod jwt;
 pub mod predicates;
